+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+---
+
+MOVE THIS SECTION TO THE CORRECT LOCATION
+
+### Misc. Updates
+
+- Model updates for the following services
+
+- Amazon Cognito Identity Provider
+ - Amazon EC2
+ - Amazon Kinesis Firehose
+ - Amazon Pinpoint
+ - Amazon Lex
+
+---
 # AWS Mobile SDK for iOS CHANGELOG
 
 ## 2.12.0